Removed Arena class
